feat: Add new field 'justificativa_da_improcedencia' to FIELDS
This change adds a new field 'justificativa_da_improcedencia' to the FIELDS dictionary in the attrs.py module. The new field has the following properties:

- Field ID: 1968
- Label: "Justificativa da Improcedência:"
- Options: ["Reclamante informou que interferência cessou", "Dados obtidos indicam que interferência não procede", "Não foi constatada portadora interferente em campo"]

This new field will be used to capture the justification for marking an issue as 'improcedente' (unfounded).

Additionally, the label for the 'gerar_plai' field has been updated from "Gerar Plai" to "Gerar Plai?".